bot/requirements.txt
httpx==0.28.1
pydantic_settings==2.11.0
pytest==8.4.2
pytest_asyncio==1.2.0
pytest_httpx==0.35.0
python-telegram-bot==22.5
